+++ mcp_tools/binary_analysis/enhanced_binary_analysis_and_exploitation.py
@@ -3,73 +3,9 @@
 from typing import Dict, Any
 
 def register_enhanced_binary_analysis_and_exploitation_tools(mcp, hexstrike_client, logger):
-    @mcp.tool()
-    def ghidra_analysis(binary: str, project_name: str = "hexstrike_analysis",
-                       script_file: str = "", analysis_timeout: int = 300,
-                       output_format: str = "xml", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Ghidra for advanced binary analysis and reverse engineering.
 
-        Args:
-            binary: Path to the binary file
-            project_name: Ghidra project name
-            script_file: Custom Ghidra script to run
-            analysis_timeout: Analysis timeout in seconds
-            output_format: Output format (xml, json)
-            additional_args: Additional Ghidra arguments
 
-        Returns:
-            Advanced binary analysis results from Ghidra
-        """
-        data = {
-            "binary": binary,
-            "project_name": project_name,
-            "script_file": script_file,
-            "analysis_timeout": analysis_timeout,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Ghidra analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/ghidra", data)
-        if result.get("success"):
-            logger.info(f"✅ Ghidra analysis completed for {binary}")
-        else:
-            logger.error(f"❌ Ghidra analysis failed for {binary}")
-        return result
 
-    @mcp.tool()
-    def pwntools_exploit(script_content: str = "", target_binary: str = "",
-                        target_host: str = "", target_port: int = 0,
-                        exploit_type: str = "local", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Pwntools for exploit development and automation.
-
-        Args:
-            script_content: Python script content using pwntools
-            target_binary: Local binary to exploit
-            target_host: Remote host to connect to
-            target_port: Remote port to connect to
-            exploit_type: Type of exploit (local, remote, format_string, rop)
-            additional_args: Additional arguments
-
-        Returns:
-            Exploit execution results
-        """
-        data = {
-            "script_content": script_content,
-            "target_binary": target_binary,
-            "target_host": target_host,
-            "target_port": target_port,
-            "exploit_type": exploit_type,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Pwntools exploit: {exploit_type}")
-        result = hexstrike_client.safe_post("api/tools/pwntools", data)
-        if result.get("success"):
-            logger.info(f"✅ Pwntools exploit completed")
-        else:
-            logger.error(f"❌ Pwntools exploit failed")
-        return result
 
     @mcp.tool()
     def one_gadget_search(libc_path: str, level: int = 1, additional_args: str = "") -> Dict[str, Any]:
@@ -124,37 +60,6 @@ def register_enhanced_binary_analysis_and_exploitation_tools(mcp, hexstrike_clie
             logger.info(f"✅ libc-database {action} completed")
         else:
             logger.error(f"❌ libc-database {action} failed")
-        return result
-
-    @mcp.tool()
-    def gdb_peda_debug(binary: str = "", commands: str = "", attach_pid: int = 0,
-                      core_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute GDB with PEDA for enhanced debugging and exploitation.
-
-        Args:
-            binary: Binary to debug
-            commands: GDB commands to execute
-            attach_pid: Process ID to attach to
-            core_file: Core dump file to analyze
-            additional_args: Additional GDB arguments
-
-        Returns:
-            Enhanced debugging results with PEDA
-        """
-        data = {
-            "binary": binary,
-            "commands": commands,
-            "attach_pid": attach_pid,
-            "core_file": core_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting GDB-PEDA analysis: {binary or f'PID {attach_pid}' or core_file}")
-        result = hexstrike_client.safe_post("api/tools/gdb-peda", data)
-        if result.get("success"):
-            logger.info(f"✅ GDB-PEDA analysis completed")
-        else:
-            logger.error(f"❌ GDB-PEDA analysis failed")
         return result
 
     @mcp.tool()
