+++ mcp_tools/exploit_framework/pwntools.py
@@ -1,0 +1,39 @@
+# mcp_tools/exploit_framework/pwntools.py
+
+from typing import Dict, Any
+
+def register_pwntools(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def pwntools_exploit(script_content: str = "", target_binary: str = "",
+                        target_host: str = "", target_port: int = 0,
+                        exploit_type: str = "local", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Pwntools for exploit development and automation.
+
+        Args:
+            script_content: Python script content using pwntools
+            target_binary: Local binary to exploit
+            target_host: Remote host to connect to
+            target_port: Remote port to connect to
+            exploit_type: Type of exploit (local, remote, format_string, rop)
+            additional_args: Additional arguments
+
+        Returns:
+            Exploit execution results
+        """
+        data = {
+            "script_content": script_content,
+            "target_binary": target_binary,
+            "target_host": target_host,
+            "target_port": target_port,
+            "exploit_type": exploit_type,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting Pwntools exploit: {exploit_type}")
+        result = hexstrike_client.safe_post("api/tools/pwntools", data)
+        if result.get("success"):
+            logger.info(f"✅ Pwntools exploit completed")
+        else:
+            logger.error(f"❌ Pwntools exploit failed")
+        return result
